Fix for failing tests - checked
